requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.0
Flask==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.24.4
openai==1.9.0
packaging==23.2
pandas==2.0.3
pipreqs==0.4.13
pyarrow==15.0.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
yarg==0.1.9
zipp==3.17.0
psycopg2-binary
dj-database-url